fix: fix error handling in log group creation
The error handling logic for creating a log group was inverted. It would
attempt to create a log group when it encountered an error *other than*
`ResourceNotFoundException`. I've corrected this to create the log group
only when it doesn't exist. I also added handling for a race condition
where another process might create the log group in the meantime.