fix: count size correctly
The size of an event was calculated from the message string only, but the
AWS API requires the size of the whole serialized event, which includes
the timestamp and some overhead.

This could lead to batches being larger than the allowed 1MB limit,
which would cause the PutLogEvents API to return an error of 400 Bad Request.

Few cosmetic issues were also fixed and error counting was added.